+++ .vscode/settings.json
@@ -7,4 +7,6 @@
     "search.exclude": {
       "**/node_modules": true,
   },
+  "remote.autoForwardPorts": false,
+  "editor.formatOnSave": true,
 }
