+++ backend/index.js
@@ -9,7 +9,7 @@ app.use(cors())
 //dotenv.config({ path: './config/.env' });
 
 //const DB_URL = process.env.MDB_URL || "mongodb://db:27018/books"
-const DB_URL = "mongodb://db:27017/books"
+const DB_URL = "mongodb://localhost:27018/books"
 console.log(DB_URL)
 app.use(bodyParser.urlencoded({ extended: true }))
