+++ frontend/src/index.css
@@ -5,7 +5,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  animation: rotation 20000s infinite linear;
+  animation: rotation 0s infinite linear;
 
 }
 
@@ -25,7 +25,7 @@ code {
 
 
 .card {
-  width: 14rem;
+  width: 18rem;
   margin-right: 1rem;
   margin-top: 1rem;
   z-index: 1;
@@ -44,47 +44,17 @@ code {
 
 .edit-button {
   position: absolute;
-  top: 17rem;
+  top: 27rem;
   right: 0rem;
   z-index: 4;
-}
-
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width:100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.6);
-  z-index: 1500;
-
-}
-
-.modal-main {
-  position: fixed;
-  background: white;
-  width: 100%;
-  height: auto;
-  top:50%;
-  left:50%;
-  transform: translate(-50%,-50%);
-  z-index: 1501;
-}
-
-.modal-backdrop {
-  display: none;
-}
-
-.display-block {
-  display: block;
-}
-
-.display-none {
-  display: none;
 }
 
 img {
   object-fit: cover;
   width: 100%;
   height: auto;
+}
+
+.card-title {
+  margin-bottom: -0.25rem;
 }
